client gets join kind

+++ Source/Core/KARphin/WarpRelay/Packets/Packet_OnPlayerConnect.hpp
@@ -33,7 +33,7 @@ namespace KARphin::WarpRelay::Netplay::Packet
 	static inline sf::Packet PackageDataIntoSFMLPacket_OnPlayerConnect(const Packet_OnPlayerConnect& data)
 	{
     sf::Packet packet;
-    packet << (int)data.kind;
+    packet << (uint8_t)data.kind;
 		packet << data.SCMVersion;
     packet << data.clientVersion;
     packet << data.nickname;
@@ -46,7 +46,7 @@ namespace KARphin::WarpRelay::Netplay::Packet
   static inline Packet_OnPlayerConnect UnpackSFMLPacket_OnPlayerConnect(sf::Packet& packet)
   {
     Packet_OnPlayerConnect data;
-    int kind = 0;
+    uint8_t kind = 0;
     packet >> kind;
     packet >> data.SCMVersion;
     packet >> data.clientVersion;
